[DEPYLER-0149] Phase 1d/1e Complete - Re-transpile and Validate Showcase Examples
Phase 1 (100% Complete) - All showcase examples re-transpiled with fixes

## Re-transpilation Results

Successfully transpiled 5/6 showcase examples:
1. binary_search.py → binary_search.rs ✅
2. calculate_sum.py → calculate_sum.rs ✅
3. classify_number.py → classify_number.rs ✅
4. contracts_example.py → contracts_example.rs ✅
5. process_config.py → process_config.rs ✅
6. annotated_example.py → ❌ (blocked by Phase 2 - dict augmented assignment)

## Validation Results

Compilation status (4/6 compile cleanly):
1. binary_search.rs - ✅ Compiles (1 warning: unnecessary parens)
2. calculate_sum.rs - ✅ Compiles (0 warnings)
3. classify_number.rs - ✅ Compiles (1 warning: unused import)
4. contracts_example.rs - ⚠️ Partial (binary_search ✅, list_sum has unrelated issues)
5. process_config.rs - ✅ Compiles (0 warnings)
6. annotated_example.rs - ❌ Does not exist

## Key Achievement

contracts_example.py binary_search function now compiles with 0 errors!

Before Phase 1: 4 type errors
After Phase 1: 0 errors ✅

This was the primary goal of Phase 1 and it is achieved!

## What Changed

All showcase examples now use:
- PEP 585 type syntax: list[int] → Vec<i32>
- Type conversion: int(x) removed (type inference)
- Consistent integers: len() → (len() as i32)

## Phase 1 Summary

Phases 1a + 1b + 1c + 1d + 1e = 100% Complete

- 1a: PEP 585 type parsing ✅
- 1b: Type conversion functions ✅
- 1c: Integer type consistency ✅
- 1d: Re-transpile showcase examples ✅
- 1e: Validate compilation ✅

## Documentation Updated

- docs/sessions/2025-10-10-technical-debt-and-planning.md
  - Added Phase 1 completion section
  - Documented validation results
  - Captured metrics and lessons learned

## Metrics

- Tests: 393 → 403 (+10)
- Transpilable: 5/6 (83%)
- Compilable: 4/6 (67%)
- binary_search errors: 4 → 0 ✅

## Next Phase

Phase 2: DEPYLER-0148 - Dict/List Augmented Assignment
- Unblocks annotated_example.py
- Estimated: 8-12 hours

Closes: DEPYLER-0149 Phases 1d, 1e
Closes: DEPYLER-0149 Phase 1 (Complete)
Progress: v3.14.0 Phase 1 - 100% COMPLETE

+++ examples/showcase/binary_search.rs
@@ -1,6 +1,6 @@
 #[doc = "Find target in sorted array, return -1 if not found."] pub fn binary_search<'a>(arr: & 'a Vec<i32>, target: i32)  -> Result<i32, Box<dyn std::error::Error>>{
     let mut left: i32  = (0) as i32;
-    let _cse_temp_0 = arr.len();
+    let _cse_temp_0  = (arr.len() as i32);
     let mut right: i32  = (_cse_temp_0 - 1) as i32;
     while left <= right {
     let mid: i32  = ({ let a = left + right;
